apply the general function
