+++ .gitignore
@@ -1,0 +1,2 @@
+#ignore all the csv files
+*.csv
