+++ task1/linear_regression/config.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+#!/usr/bin/python
+
+HOUSE_DS = "housing_dataset.csv"
+TITANIC_DS = "titanic dataset.csv"
+CHUNKS = [100, 500, 1000, 5000, 10000, 50000, 100000, 500000, 1000000, 5000000, 10000000, 50000000, 100000000]
+SUM_WO_NOISE_DS = "The SUM dataset, without noise.csv"
+SUM_WI_NOISE_DS = "The SUM dataset, with noise.csv"
